Preparation for larger changes

+++ src/components/fitness/gpu/kernels.h
@@ -13,9 +13,9 @@ namespace eic {
 
 
 __global__
-void populationFitness (__uint8_t *g_target, unsigned int width, unsigned int height, float *g_population,
+void populationFitness (__uint8_t *g_target, unsigned int width, unsigned int height, int *g_population,
                         unsigned int population_size, unsigned int chromozome_length, float *g_out_fitness,
-                        float *s_canvas);
+                        int *s_canvas);
 
 
 }
